+++ src/backend/marsha/core/tests/api/video/test_transcript_source.py
@@ -4,8 +4,6 @@ from datetime import datetime, timezone
 from http import HTTPStatus
 
 from django_peertube_runner_connector.factories import RunnerJobFactory
-import requests
-import responses
 from rest_framework.test import APITestCase
 
 from marsha.core import factories
@@ -26,30 +24,21 @@ class VideoTranscriptSourceAPITest(APITestCase):
         )
         self.url = f"/api/videos/{self.video.pk}/transcript-source/"
         serializer = VideoSerializer(self.video)
-        self.video_url = serializer.data.get("urls").get("mp4").get(144)
+        self.video_url = serializer.data.get("urls").get("mp4").get(720)
         runner_job = RunnerJobFactory()
         self.runner_job_data = {
             "runnerToken": runner_job.runner.runnerToken,
             "jobToken": runner_job.processingJobToken,
         }
 
-    @responses.activate
     def test_api_video_transcript_source(self):
-        """The API should return the video source."""
-        responses.get(
-            url=self.video_url,
-            body=b"video content",
-            status=200,
-            content_type="video/mp4",
-        )
+        """The API should redirect to the video url."""
 
         response = self.client.post(self.url, data=self.runner_job_data)
 
-        self.assertEqual(response.status_code, HTTPStatus.OK)
-        self.assertEqual(response["Content-Type"], "video/mp4")
-        self.assertEqual(b"".join(response.streaming_content), b"video content")
+        self.assertEqual(response.status_code, HTTPStatus.MOVED_PERMANENTLY)
+        self.assertEqual(response["Location"], self.video_url)
 
-    @responses.activate
     def test_api_video_transcript_source_missing_video_url(self):
         """The API should return a 404 if the video source is not available."""
         self.video.resolutions = []
@@ -63,28 +52,8 @@ class VideoTranscriptSourceAPITest(APITestCase):
             status_code=HTTPStatus.NOT_FOUND,
         )
 
-    @responses.activate
-    def test_api_video_transcript_source_request_failure(self):
-        """The API should return a 400 if the request to the video source fails."""
-        responses.get(
-            url=self.video_url,
-            body=requests.RequestException(),
-            status=500,
-        )
-
-        response = self.client.post(self.url, data=self.runner_job_data)
-
-        self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
-
-    @responses.activate
     def test_api_video_transcript_source_no_runner(self):
         """The API should return a 403 if the request is not from a runner job."""
-        responses.get(
-            url=self.video_url,
-            body=b"video content",
-            status=200,
-            content_type="video/mp4",
-        )
         data = {
             "runnerToken": "unknown-runner-token",
             "jobToken": "unknown-job-token",